Refactor CalculatePosition as extension method
